HTML + TeX from script
Reorganized the structure of the 'highlighter'. Introduced new tags for marking the start of a non-code environment.
HTML output works as before, but css file has to be recreated.
TeX file is organized into environments, but the proper setup for the new environments has to be created as well.
The math operators ^, _ are not handled in every case.